Add client APIs integ test
